Clean up isgx_user.h
It is recommended for new upstream drivers that the same binary format
would be used for 32-bit and 64-bit driver so that no compat layer is
needed. This changes address fields to __u64.

This commit moves architectural structures isgx_sigstruct and
isgx_einittoken to isgx_arch.h. It is better for the user space to
define its own versions of these structures so that it can update them
without relaying to a particular kernel release.

There is another less technical reason for dropping as much as
structures possible from this header. Every struct that we have
is a target for feedback so there might be discussion about naming
or whatever. Thus less we have structs and contants, less likely
we are to have new change requests to this header file.

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@linux.intel.com>