isgx: fix custom return values
The custom return values use the base 0xc00000000, which is a negative
value in twos complement form since the highest bit is up. See this
Python example:

>>> import ctypes
>>> ctypes.c_int32(0xc0000000)
c_int(-1073741824)

However, the man-page of ioctl says that:

"Usually, on success zero is returned.  A few ioctl() requests use the
return value as an output parameter and return a nonnegative value on
success."

It is probably better to use the base 0x40000000 instead although
man-page does not explicitly forbird the use of negative values.

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@linux.intel.com>