sgx_main: clean up sgx_drv_probe()
It is unnecesary to implement the same logic in two places -
sgx_init_platform() and sgx_drv_probe().

One bad example is that the evaluation of max enclave size is wrong
in sgx_drv_probe(). Due to the calling order, the final result is
still correct.

Therefore, move all sanity checks into the internal function
sgx_init_platform() which needs to know all details of sgx.

Signed-off-by: Jia Zhang <qianyue.zj@alibaba-inc.com>