Fix incorrect entry return on sgx_do_fault
Linux version 4.20+ returns VM_FAULT_NOPAGE on vmf_insert_pfn()
which leads to sgx_do_fault() incorrectly treating this value
as an error and propagating it as an address to other functions.

In particular, this bug leads to sgx_vma_access_word() working
on a meaningless encl_page which ends in a kernel oops (due to
incorrect memory access).

This commit overwrites return value from VM_FAULT_NOPAGE to zero
on happy path in sgx_do_fault().