Performance improvements
The epc free page list spinlock was identified as a bottleneck
during eviction stress. To overcome this issue, an atomic simple
list data structure was implemented.
The wake_up api is used to trigger calls to ksgxswap. This method
uses a spinlock. It has been identified as a bottleneck as well.
A simple strategy using a boolean flag was adopted in order to
avoid most of those calls.
Removing directory inker2ext which is not relevant anymore.

Signed-off-by: Ayoun Serge <serge.ayoun@intel.com>